Ajuste a controlador de usuario e imagen de perfil por defecto

+++ app/Http/Controllers/UserController.php
@@ -28,10 +28,12 @@ class UserController extends Controller
         $user->avatar = $avatar;
         $user->save();
 
-        Storage::disk('public')->delete($oldImage);
+        if(strpos($oldImage, 'default.jpg') ===false ){
+          Storage::disk('public')->delete($oldImage);
+        }
 
         return view('profile', [
-          'user' => Auth::user()
+          'user' => $user
         ]);
       }
     }
